requirements.txt
functions-framework==3.5.0
google-cloud-secret-manager==2.20.0
google-cloud-bigquery==3.19.0
google-auth==2.29.0
google-api-python-client==2.127.0
requests==2.31.0
sendgrid==6.11.0
twilio==9.0.4
